% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/df2json.R
\name{df2json}
\alias{df2json}
\title{Converts data frame to hierarchical JSON}
\usage{
df2json(d)
}
\arguments{
\item{d}{A length 1 character vector contain the name of the data frame to be converted}
}
\value{
Hierarchical JSON object.
}
\description{
Converts data frame to hierarchical JSON
}
\details{
Function assumes the order of the columns is the hierarchical order
}
\examples{
data(mtcars)
nd = mtcars \%>\%
  dplyr::mutate(mpg = cut(mpg,c(floor(min(.$mpg)),15,20,25,ceiling(max(.$mpg))),include.lowest = T)) \%>\%
  dplyr::mutate(am = ifelse(am == 0, "Automatic","Manual")) \%>\%
  dplyr::select(am,gear,mpg,carb,cyl)
json = df2json("nd")
json
}

