% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dndTree.R
\name{renderDndTree}
\alias{renderDndTree}
\title{Widget render function for use in Shiny}
\usage{
renderDndTree(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

